requirements.txt
openai
python-dotenv
langchain-openai
langchain-ollama
langchain-google-genai
langchain-anthropic
langchain-huggingface
transformers
torch